Move test module out of app directory.
