+++ ui/src/app/app.component.spec.ts
@@ -1,5 +1,5 @@
 import { TestBed, async } from '@angular/core/testing';
-import { TestModule } from './test.module';
+import { TestModule } from './../test.module';
 
 import { AppComponent } from './app.component';
