+++ ui/src/app/reader/reader.component.spec.ts
@@ -1,5 +1,5 @@
 import { async, ComponentFixture, TestBed } from '@angular/core/testing';
-import { TestModule } from './../test.module';
+import { TestModule } from './../../test.module';
 
 import { ReaderComponent } from './reader.component';
