+++ ui/src/app/volumes/volumes.component.html
@@ -1,0 +1,3 @@
+<p>
+  volumes works!
+</p>
