+++ ui/src/app/volumes/volumes.component.ts
@@ -1,0 +1,15 @@
+import { Component, OnInit } from '@angular/core';
+
+@Component({
+  selector: 'app-volumes',
+  templateUrl: './volumes.component.html',
+  styleUrls: ['./volumes.component.scss']
+})
+export class VolumesComponent implements OnInit {
+
+  constructor() { }
+
+  ngOnInit() {
+  }
+
+}
